@@ -7,6 +7,8 @@ import {
 
 import moment from 'moment/min/moment-with-locales.min';
 
+import { isSameDay, isSameUser, warnDeprecated } from './utils';
+
 export default class Day extends React.Component {
   dayFormat() {
     if( this.props.dayFormat ) {
@@ -18,7 +20,7 @@ export default class Day extends React.Component {
   }
 
   render() {
-    if (!this.props.isSameDay(this.props.currentMessage, this.props.previousMessage)) {
+    if (!isSameDay(this.props.currentMessage, this.props.previousMessage)) {
       return (
         <View style={[styles.container, this.props.containerStyle]}>
           <View style={[styles.wrapper, this.props.wrapperStyle]}>
@@ -61,7 +63,6 @@ Day.contextTypes = {
 };
 
 Day.defaultProps = {
-  isSameDay: () => {},
   currentMessage: {
     // TODO test if crash when createdAt === null
     createdAt: null,
@@ -78,14 +79,19 @@ Day.defaultProps = {
   containerStyle: {},
   wrapperStyle: {},
   textStyle: {},
+  //TODO: remove in next major release
+  isSameDay: warnDeprecated(isSameDay),
+  isSameUser: warnDeprecated(isSameUser),
 };
 
 Day.propTypes = {
-  isSameDay: React.PropTypes.func,
   currentMessage: React.PropTypes.object,
   previousMessage: React.PropTypes.object,
   dayFormat: React.PropTypes.object,
   containerStyle: View.propTypes.style,
   wrapperStyle: View.propTypes.style,
   textStyle: Text.propTypes.style,
+  //TODO: remove in next major release
+  isSameDay: React.PropTypes.func,
+  isSameUser: React.PropTypes.func,
 };
