@@ -140,11 +140,13 @@ export default class MessageContainer extends React.Component {
       <View ref='container' style={{flex:1}}>
         <ListView
           enableEmptySections={true}
-          keyboardShouldPersistTaps={true}
           automaticallyAdjustContentInsets={false}
           initialListSize={20}
           pageSize={20}
 
+          {...this.props.listViewProps}
+
+          enableEmptySections={true}
           dataSource={this.state.dataSource}
 
           renderRow={this.renderRow}
@@ -162,6 +164,7 @@ MessageContainer.defaultProps = {
   user: {},
   renderFooter: null,
   renderMessage: null,
+  listViewProps: {},
   onLoadEarlier: () => {
   },
 };
@@ -172,4 +175,5 @@ MessageContainer.propTypes = {
   renderFooter: React.PropTypes.func,
   renderMessage: React.PropTypes.func,
   onLoadEarlier: React.PropTypes.func,
+  listViewProps: React.PropTypes.object,
 };
