@@ -3,16 +3,28 @@ import {
   Image,
   StyleSheet,
   View,
+  Dimensions,
 } from 'react-native';
+import Lightbox from 'react-native-lightbox';
 
 export default class MessageImage extends React.Component {
   render() {
+    const { width, height } = Dimensions.get('window');
+
     return (
       <View style={[styles.container, this.props.containerStyle]}>
-        <Image
-          style={[styles.image, this.props.imageStyle]}
-          source={{uri: this.props.currentMessage.image}}
-        />
+        <Lightbox
+          activeProps={{
+            style: [styles.imageActive, { width, height }],
+          }}
+          {...this.props.lightboxProps}
+        >
+          <Image
+            {...this.props.imageProps}
+            style={[styles.image, this.props.imageStyle]}
+            source={{uri: this.props.currentMessage.image}}
+          />
+        </Lightbox>
       </View>
     );
   }
@@ -28,6 +40,9 @@ const styles = StyleSheet.create({
     margin: 3,
     resizeMode: 'cover',
   },
+  imageActive: {
+    resizeMode: 'contain',
+  },
 });
 
 MessageImage.defaultProps = {
@@ -36,10 +51,14 @@ MessageImage.defaultProps = {
   },
   containerStyle: {},
   imageStyle: {},
+  imageProps: {},
+  lightboxProps: {},
 };
 
 MessageImage.propTypes = {
   currentMessage: React.PropTypes.object,
   containerStyle: View.propTypes.style,
   imageStyle: Image.propTypes.style,
+  imageProps: React.PropTypes.object,
+  lightboxProps: React.PropTypes.object,
 };
